requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
openai==1.12.0
apscheduler==3.10.4
python-multipart==0.0.6
logging==0.5.1.2